Changed microservice landscape to support reactive streams

+++ microservices/recommendation-service/src/main/java/com/aw/micro/recommendation_service/service/RecommendationServiceImpl.java
@@ -11,10 +11,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.logging.Level;
 
 @RestController
 public class RecommendationServiceImpl implements RecommendationService {
@@ -33,38 +34,48 @@ public class RecommendationServiceImpl implements RecommendationService {
   }
 
   @Override
-  public List<Recommendation> getRecommendations(int productId) {
+  public Flux<Recommendation> getRecommendations(int productId) {
 
     if (productId < 1) {
       throw new InvalidInputException("Invalid productId: " + productId);
     }
+    LOG.info("Will get recommendations for product with id={}", productId);
 
-    List<RecommendationEntity> entityList=repository.findByProductId(productId);
-    List<Recommendation> list = mapper.entityListToApiList(entityList);
-    list.forEach(e->e.setServiceAddress(serviceUtil.getServiceAddress()));
+    return repository.findByProductId(productId)
+            .log(LOG.getName(), Level.FINE)
+            .map(mapper::entityToApi)
+            .map(this::setServiceAddress);
 
-    LOG.debug("getRecommendations: response size: {}", list.size());
 
-
-    return list;
   }
 
   @Override
-  public Recommendation createRecommendation(Recommendation body) {
-    try{
-      RecommendationEntity recommendationEntity = mapper.apiToEntity(body);
-      RecommendationEntity newEntity = repository.save(recommendationEntity);
-
-      LOG.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(), body.getRecommendationId());
-      return mapper.entityToApi(newEntity);
-    }catch (DuplicateKeyException dke){
-      throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId());
+  public Mono<Recommendation> createRecommendation(Recommendation body) {
+    if (body.getProductId() < 1) {
+      throw new InvalidInputException("Invalid productId: " + body.getProductId());
     }
+    RecommendationEntity entity = mapper.apiToEntity(body);
+
+    return repository.save(entity)
+            .log(LOG.getName(),Level.FINE)
+            .onErrorMap(DuplicateKeyException.class,
+                    ex -> new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId()))
+            .map(e -> mapper.entityToApi(e));
   }
 
   @Override
-  public void deleteRecommendation(int productId) {
+  public Mono<Void> deleteRecommendation(int productId) {
+    if (productId < 1) {
+      throw new InvalidInputException("Invalid productId: " + productId);
+    }
+
     LOG.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}", productId);
-    repository.deleteAll(repository.findByProductId(productId));
+    return repository.deleteAll(repository.findByProductId(productId));
+  }
+
+
+  private Recommendation setServiceAddress(Recommendation e) {
+    e.setServiceAddress(serviceUtil.getServiceAddress());
+    return e;
   }
 }
